requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
pydantic
pydantic-settings
httpx
